% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AustrianWages.R
\docType{data}
\name{AustrianWages}
\alias{AustrianWages}
\title{AustrianWages}
\format{
A \code{list} with one element, which is:
\describe{
\item{\code{data}}{A list with 9402 MTS.}
}
}
\usage{
data(AustrianWages)
}
\description{
Ordinal time series (OTS) of yearly categories of salaries for different
Austrian employees
}
\details{
Each element in \code{data} is an ordinal time series
containing 6 states (yearly categorized wages). 9402 Austrian workers
considered. The series exhibit individual lengths ranging from 2 to 32 years with the median length being equal to 22.
For more information, see \insertCite{lopez2023submitted;textual}{otsfeatures}.
}
\references{
{

  \insertRef{lopez2023submitted}{otsfeatures}

}
}
\keyword{datasets}
