% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ots_plot.R
\encoding{UTF-8}
\name{ots_plot}
\alias{ots_plot}
\title{Constructs an ordinal time series plot}
\usage{
ots_plot(series, states, title = "Time series plot", labels = NULL)
}
\arguments{
\item{series}{An OTS.}

\item{states}{A numerical vector containing the corresponding
states.}

\item{title}{The title of the graph.}

\item{labels}{The labels of the graph.}
}
\value{
The ordinal time series plot.
}
\description{
\code{ots_plot} constructs an ordinal time series plot
}
\details{
Constructs an ordinal time series plot for a given OTS.
}
\examples{
ordinal_time_series_plot <- ots_plot(series = AustrianWages$data[[100]],,
states = 0 : 5) # Constructs an ordinal
# time series plot for one series in
# dataset AustrianWages
}
\references{
{

  \insertRef{weiss2018introduction}{otsfeatures}

}
}
\author{
Ángel López-Oriona, José A. Vilar
}
